+++ src/GridPanel.cpp
@@ -52,9 +52,9 @@ void GridPanel::resize(const sf::VideoMode &newSize) {
     // set the outline percentage (of width)
     const float outline = 0.05f * size.x;
 
-    for (auto i = 0; i < nRows; ++i) {
+    for (auto i = 0u; i < nRows; ++i) {
         rects.emplace_back(RowVector{});
-        for (auto j = 0; j < nCols; ++j) {
+        for (auto j = 0u; j < nCols; ++j) {
             sf::RectangleShape rect;
             sf::Vector2f pos(j * size.x, i * size.y);
 
@@ -190,8 +190,8 @@ void GridPanel::kbKeyRelHandle(sf::Event &e) {
 GridPanel::Idx GridPanel::rectToIdx(GridPanel::Rect *rect) {
     auto& shape = rect->first;
     Idx index;
-    index.second = shape.getPosition().x / shape.getSize().x;
-    index.first  = shape.getPosition().y / shape.getSize().y;
+    index.second = static_cast<unsigned>(shape.getPosition().x / shape.getSize().x);
+    index.first  = static_cast<unsigned>(shape.getPosition().y / shape.getSize().y);
     return index;
 }
 
@@ -210,13 +210,13 @@ GridPanel::Rect* GridPanel::findNeighbour(GridPanel::Rect *src, int dir) {
             if (col == (nCols - 1))
                 ret =  nullptr;
             else
-                ret =  &rects[row][col + 1];
+                ret =  &rects[row][(unsigned long long)col + 1];
             break;
         case 2:
             if (row == (nRows - 1))
                 ret = nullptr;
             else
-                ret = &rects[row + 1][col];
+                ret = &rects[(unsigned long long)row + 1][col];
             break;
         case 3:
             if (col == 0)
